pyproject.toml
[tool.poetry]
name = "nst"
version = "0.1.0"
description = ""
authors = ["alessandro.paticchio <alessandropaticchio@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
torch = "^1.12.0"
Pillow = "^9.2.0"
numpy = "^1.23.1"
torchvision = "^0.13.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
